refactor: move get_access_token, make_request, and BaseIntegrationTest to eox_core/test_utils
